refactor(utils): update norm_order parameter to double for qalsh config regularization

+++ src/utils.h
@@ -16,7 +16,7 @@ class Utils {
     static std::vector<Point> LoadPointsFromFile(const std::filesystem::path &file_path, unsigned int num_points,
                                                  unsigned int num_dimensions);
     static Point ReadPoint(std::ifstream &ifs, unsigned int num_dimensions, unsigned int point_id);
-    static void RegularizeQalshConfig(QalshConfig &config, unsigned int num_points, unsigned int norm_order);
+    static void RegularizeQalshConfig(QalshConfig &config, unsigned int num_points, double norm_order);
     static void SaveQalshConfig(QalshConfig &config, const std::filesystem::path &file_path);
     static QalshConfig LoadQalshConfig(const std::filesystem::path &file_path);
     static unsigned int SampleFromWeights(const std::vector<double> &weights);
